backend/diagnosticore-service/path_foundation_experiment/requirements.txt
huggingface-hub>=0.26.0,<1.0.0
numpy>=1.26.0
scikit-learn>=1.3.0
Pillow>=10.0.0
tqdm>=4.66.0
joblib>=1.3.0
tensorflow>=2.16.0; python_version < "3.13"
tf-keras>=2.16.0; python_version < "3.13"
